Build navigational elements: breadcrumbs
Fix the timestamp issue (finally) - there was an issue with DST. Timestamping
now achieved by means of the native WP `current_time()` function.
This takes account of the local blog time settings, so is most appropriate
for inclusion in a distributed plugin.

I sill can't see the DB value for gmt_offset in the options table, though the
value IS being applied!