Fix filter blank space in table
This is achieved by placing the table in a containing div.
On page load, the table height is calculated using jQuery.
This height is then applied to the table containing div.
When the table collapses as elements are dynamically hidden, the
container maintains it's height. This prevents jumping on the page
when table elements are hidden.